+++ util/cache_test.cc
@@ -16,6 +16,7 @@
 #include <vector>
 #include "util/clock_cache.h"
 #include "util/coding.h"
+#include "util/lru_cache.h"
 #include "util/string_util.h"
 #include "util/testharness.h"
 
@@ -73,6 +74,17 @@ class CacheTest : public testing::TestWithParam<std::string> {
   }
 
   ~CacheTest() {
+  }
+
+  std::shared_ptr<Cache> NewCache(size_t capacity) {
+    auto type = GetParam();
+    if (type == kLRU) {
+      return NewLRUCache(capacity);
+    }
+    if (type == kClock) {
+      return NewClockCache(capacity);
+    }
+    return nullptr;
   }
 
   std::shared_ptr<Cache> NewCache(size_t capacity, int num_shard_bits,
@@ -626,6 +638,22 @@ TEST_P(CacheTest, ApplyToAllCacheEntiresTest) {
   std::sort(inserted.begin(), inserted.end());
   std::sort(callback_state.begin(), callback_state.end());
   ASSERT_TRUE(inserted == callback_state);
+}
+
+TEST_P(CacheTest, DefaultShardBits) {
+  // test1: set the flag to false. Insert more keys than capacity. See if they
+  // all go through.
+  std::shared_ptr<Cache> cache = NewCache(16 * 1024L * 1024L);
+  ShardedCache* sc = dynamic_cast<ShardedCache*>(cache.get());
+  ASSERT_EQ(5, sc->GetNumShardBits());
+
+  cache = NewLRUCache(511 * 1024L, -1, true);
+  sc = dynamic_cast<ShardedCache*>(cache.get());
+  ASSERT_EQ(0, sc->GetNumShardBits());
+
+  cache = NewLRUCache(1024L * 1024L * 1024L, -1, true);
+  sc = dynamic_cast<ShardedCache*>(cache.get());
+  ASSERT_EQ(6, sc->GetNumShardBits());
 }
 
 #ifdef SUPPORT_CLOCK_CACHE